% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf_calculations.R
\name{ddf_calculations}
\alias{ddf_calculations}
\title{Historical and difference DDF calculation}
\usage{
ddf_calculations(ddf_dt)
}
\arguments{
\item{ddf_dt}{A processed dataset with historical RCP}
}
\value{
A final dataset with all the data
}
\description{
Historical and difference DDF calculation
}
