% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf_precipitations.R
\name{ddf_precipitations}
\alias{ddf_precipitations}
\title{DDF for precipitation data}
\usage{
ddf_precipitations(dt)
}
\arguments{
\item{dt}{A dataset with certain params as p(precipitation kg/s), date and SID(filenames)}
}
\value{
A preprocessed dataset with new columns
}
\description{
DDF for precipitation data
}
