% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf_temperature.R
\name{ddf_temperature}
\alias{ddf_temperature}
\title{DDF for temperature data}
\usage{
ddf_temperature(dt)
}
\arguments{
\item{dt}{A dataset with certain params as TAS(temperature), DTM(date) and SID(filenames)}
}
\value{
A dataset with new columns
}
\description{
DDF for temperature data
}
