% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_combine_raw.R
\name{read_and_combine_raw}
\alias{read_and_combine_raw}
\title{Data reading function}
\usage{
read_and_combine_raw(dir, na.rm = FALSE)
}
\arguments{
\item{dir}{dirrectory with all the raw data files}

\item{na.rm}{FALSE by default}
}
\value{
A one combined dataset with filename and data
}
\description{
Data reading function
}
\examples{

}
