% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spatial_data.R
\name{read_spatial_data}
\alias{read_spatial_data}
\title{Preprocessing spatial dataset}
\usage{
read_spatial_data(dir, ext)
}
\arguments{
\item{dir}{A directory with .nc files}

\item{ext}{Extention for the area}
}
\value{
A dataset with cells ID as columnnames and precipitation data
}
\description{
Preprocessing spatial dataset
}
